Refix merge from pull request #52

+++ src/services/api/Product.js
@@ -122,7 +122,7 @@ const Product = {
       throw new APIError(responseJSON.message || responseJSON.error);
     }
 
-    if (oldImage) {
+    if (oldImage.startsWith('https://firebasestorage.googleapis.com/')) {
       await Storage.deleteProductImage(oldImage);
     }
 
@@ -145,7 +145,9 @@ const Product = {
       throw new APIError(responseJSON.message || responseJSON.error);
     }
 
-    await Storage.deleteProductImage(image);
+    if (image.startsWith('https://firebasestorage.googleapis.com/')) {
+      await Storage.deleteProductImage(image);
+    }
 
     return responseJSON.data;
   },
